refactor: Add nested classes in new files.
