+++ apis/Google.Cloud.PubSub.V1/Google.Cloud.PubSub.V1/Google.Cloud.PubSub.V1.csproj
@@ -13,4 +13,18 @@
     <PackageReference Include="Google.Cloud.Iam.V1" Version="[3.0.0, 4.0.0)" />
     <PackageReference Include="Grpc.Core" Version="[2.46.5, 3.0.0)" PrivateAssets="None" Condition="'$(TargetFramework)'=='net462'" />
   </ItemGroup>
+  <ItemGroup>
+      <Compile Update="PublisherClient.*.cs">
+          <DependentUpon>PublisherClient.cs</DependentUpon>
+      </Compile>
+      <Compile Update="PublisherClientImpl.*.cs">
+          <DependentUpon>PublisherClientImpl.cs</DependentUpon>
+      </Compile>
+      <Compile Update="SubscriberClient.*.cs">
+          <DependentUpon>SubscriberClient.cs</DependentUpon>
+      </Compile>
+      <Compile Update="SubscriberClientImpl.*.cs">
+          <DependentUpon>SubscriberClientImpl.cs</DependentUpon>
+      </Compile>
+  </ItemGroup>
 </Project>
